feat(note): add todo list extension, separate createCommandsPlugin into a utility file
