+++ package.json
@@ -65,6 +65,7 @@
         "@tiptap/extension-heading": "3.11.0",
         "@tiptap/extension-highlight": "3.11.0",
         "@tiptap/extension-image": "3.11.0",
+        "@tiptap/extension-list": "3.11.0",
         "@tiptap/extension-placeholder": "3.11.0",
         "@tiptap/extension-table": "3.11.0",
         "@tiptap/extension-text-align": "3.11.0",
