+++ src/app/[locale]/topics/[topicId]/(topic)/components/note/style.css
@@ -9,7 +9,7 @@
 }
 
 .tiptap p.is-editor-empty:first-child::before {
-    color: #aaa;
+    @apply text-muted-foreground;
     content: attr(data-placeholder);
     float: left;
     height: 0;
@@ -108,4 +108,57 @@
 .tiptap pre code {
     display: block;
     white-space: pre;
+}
+
+.tiptap {
+    :first-child {
+        margin-top: 0;
+    }
+
+    /* List styles */
+    ul,
+    ol {
+        padding: 0 1rem;
+        margin: 1.25rem 1rem 1.25rem 0.4rem;
+
+        li p {
+            margin-top: 0.25em;
+            margin-bottom: 0.25em;
+        }
+    }
+
+    /* Task list specific styles */
+    ul[data-type='taskList'] {
+        list-style: none;
+        margin-left: 0;
+        padding: 0;
+
+        li {
+            align-items: flex-start;
+            display: flex;
+
+            > label {
+                flex: 0 0 auto;
+                margin-right: 0.5rem;
+                user-select: none;
+            }
+
+            > div {
+                flex: 1 1 auto;
+            }
+        }
+
+        li[data-checked='true'] > div > p:first-of-type {
+            text-decoration: line-through;
+            color: #888;
+        }
+
+        input[type='checkbox'] {
+            cursor: pointer;
+        }
+
+        ul[data-type='taskList'] {
+            margin: 0;
+        }
+    }
 }
